+++ Runtime/Unity.Serialization/Extensions.cs
@@ -1,5 +1,3 @@
-using System.IO;
-
 namespace Unity.Serialization.Json
 {
     static class StringExtensions
@@ -7,26 +5,6 @@ namespace Unity.Serialization.Json
         internal static string ToForwardSlash(this string value)
         {
             return value.Replace('\\', '/');
-        }
-    }
-
-    static class DirectoryInfoExtensions
-    {
-        internal static string GetRelativePath(this DirectoryInfo directoryInfo)
-        {
-            var relativePath = new DirectoryInfo(".").FullName.ToForwardSlash();
-            var path = directoryInfo.FullName.ToForwardSlash();
-            return path.StartsWith(relativePath) ? path.Substring(relativePath.Length).TrimStart('/') : path;
-        }
-    }
-
-    static class FileInfoExtensions
-    {
-        internal static string GetRelativePath(this FileInfo fileInfo)
-        {
-            var relativePath = new DirectoryInfo(".").FullName.ToForwardSlash();
-            var path = fileInfo.FullName.ToForwardSlash();
-            return path.StartsWith(relativePath) ? path.Substring(relativePath.Length).TrimStart('/') : path;
         }
     }
 }